Implement node cluster in order to speed up of query optimization
